Various bugs. Added axon elongation equation
Various experiments with WindowsApi

+++ modeling/neuron.cpp
@@ -70,7 +70,7 @@ int Neuron::addConnection(int growthConeId, Neuron* neuron) {
 	delete [] tmpConnections;
 
 	connections[numberOfConnections - 1].neuron = neuron;
-	connections[numberOfConnections - 1].delay  = axons->getGrowthConeDistance(growthConeId);
+	connections[numberOfConnections - 1].delay  = (int)axons->getGrowthConeDistance(growthConeId);
 	
 #ifdef CONNECTIONTRACES
 	TRACE("neuron", "Neuron id %d now has new connection with delay %d. The number of connections: %d\n", NeuronId, connections[numberOfConnections - 1].delay,numberOfConnections);
